GPS is now optional: request GPS access in TrackRecordingActivity.
Fixes #1061.
