+++ src/main/java/de/dennisguse/opentracks/util/PermissionUtils.java
@@ -1,0 +1,18 @@
+package de.dennisguse.opentracks.util;
+
+import android.Manifest;
+import android.content.Context;
+import android.content.pm.PackageManager;
+
+import androidx.core.content.ContextCompat;
+
+public class PermissionUtils {
+
+    private PermissionUtils() {
+    }
+
+    public static boolean hasGPSPermission(Context context) {
+        return ContextCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED
+                && ContextCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED;
+    }
+}
